requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
bitstring==3.1.9
cachelib==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
colormath==3.0.0
Flask==2.0.1
Flask-Session==0.4.0
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
grpcio==1.39.0
h11==0.12.0
idna==3.2
ifaddr==0.1.7
itsdangerous==2.0.1
Jinja2==3.0.1
lifxlan==1.2.7
MarkupSafe==2.0.1
multidict==5.1.0
networkx==2.6.2
numpy==1.21.1
packaging==21.0
-e git+ssh://git@github.com/grahamhome/pifx.git@1f84d7d13dbbcbdd4530bdf4c7699a8c317a657b#egg=pifx
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-futures==1.0.0
rsa==4.7.2
six==1.16.0
spotipy==2.18.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
Werkzeug==2.0.1
yarl==1.6.3
